requirements/base.txt
Django>=3.0,<4.0
django-modeltrans<0.8
